Clean up the crud.

+++ Triangle.java
@@ -4,9 +4,9 @@ public class Triangle extends Shape {
 	private float base;
 	private float height;
 	// i want my default triangle to have area equal to 1
-	private static int INITIAL_BASE = 1;
-	private static int INITIAL_HEIGHT = 2;
-	private static int counter = 0;
+	private static float INITIAL_BASE = 1;
+	private static float INITIAL_HEIGHT = 2;
+	private static float counter = 0;
 
 	
 	public Triangle() {
@@ -20,7 +20,7 @@ public class Triangle extends Shape {
 		counter++;
 	}
 	
-	public static int getCounter() {
+	public static float getCounter() {
 		return counter;
 	}
 
@@ -40,13 +40,13 @@ public class Triangle extends Shape {
 	public float getBase() {
 		return base;
 	}
-	public void setBase(int base) {
+	public void setBase(float base) {
 		this.base = base;
 	}
 	public float getHeight() {
 		return height;
 	}
-	public void setHeight(int height) {
+	public void setHeight(float height) {
 		this.height = height;
 	}
 	public float getArea() {
